Move Response extensions to OkHttpExtensions.kt
